using string to tupels

+++ exercises/Understanding/Java_Konstruktors2/Account_Balance/Account.java
@@ -10,7 +10,7 @@ public class Account {
     public void setbalance(double balance){
         this.balance = balance;
     }
-    public void deposit(int balance){
+    public void deposit(double balance){
         this.balance = this.balance + balance;
     }
